requirements.txt
anyio==3.7.0
assemblyai==0.11.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
Flask==2.3.2
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyAudio==0.2.13
pycparser==2.21
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
typing_extensions==4.5.0
websockets==11.0.3
Werkzeug==2.3.6
zope.event==5.0
zope.interface==6.0
